app/requirements.txt
docling==2.65.0
docling_haystack==0.1.1
haystack-ai==2.21.0
pgvector-haystack==5.4.0
ollama-haystack==5.3.0
nltk==3.9.2
numpy==2.2.6
opencv_python==4.12.0.88
pdf2image==1.17.0
Pillow==11.3.0
python_Levenshtein==0.27.3
streamlit==1.52.2
onnxruntime==1.23.2
sentence_transformers==5.2.0